+++ src/main/java/org/opensearch/knn/profiler/Computation.java
@@ -5,17 +5,20 @@
 
 package org.opensearch.knn.profiler;
 
+import org.apache.commons.math3.stat.descriptive.StatisticalSummary;
+
 /**
  * Interface defining computation operations for k-Nearest Neighbor (KNN) calculations.
  * This interface provides methods to perform numerical computations on floating point values
  * used in KNN algorithm implementations.
  */
+
 public interface Computation {
     /**
-     * Performs a computation operation on a vector of float values.
+     * Performs a computation operation on statistical summary data.
      *
-     * @param perDimensionVector Array containing values for each dimension
-     * @return An array of float values representing the computation result
+     * @param stats StatisticalSummary containing the aggregated data
+     * @return The result of the computation
      */
-    float[] compute(float[] perDimensionVector);
+    double compute(StatisticalSummary stats);
 }